Add mixin for view, this allows reuse code based on unique definition
