+++ src/lang/ADempiere/en.js
@@ -170,6 +170,11 @@ export default {
       allWindows: ' and all Windows'
     }
   },
+  grid: {
+    recordAccess: {
+      accessGranted: 'Access granted for roles'
+    }
+  },
   views: {
     browser: 'Smart Browser',
     smartBrowser: 'Smart Browser',
